news_aggregator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.1.4
pyarrow==14.0.2
pyyaml==6.0.1
requests==2.31.0
fastparquet